% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_qPCR.R
\name{analyze_qPCR}
\alias{analyze_qPCR}
\title{Analyze qPCR data}
\usage{
analyze_qPCR(
  path_to_qPCR_res,
  path_to_group_label,
  out_dir_path,
  controls = c("18S", "GAPDH", "RPL13A"),
  plot_all = TRUE,
  plot_internal_controls = FALSE,
  plot_individual = FALSE,
  save_final_table = TRUE,
  file_prefix = "",
  dpi = 150
)
}
\arguments{
\item{path_to_qPCR_res}{path to a csv file produced by QuantStudio.}

\item{path_to_group_label}{path to a csv file including meta data.}

\item{out_dir_path}{path to an output directory.}

\item{controls}{list of internal controls.
c("18S", "GAPDH", "RPL13A") (default)}

\item{plot_all, plot_internal_controls, plot_individual, save_final_table}{A boolean.}

\item{file_prefix}{prefix for files.}

\item{dpi}{dpi for graphs.}
}
\description{
\code{analyze_qPCR()} analyzes a csv file produced by QuantStudio based on
the ddCT method. This function implements the method described in
this reference.
Reference https://www.cell.com/trends/biotechnology/fulltext/S0167-7799(18)30342-1
}
