% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{build_expression_table}
\alias{build_expression_table}
\title{Build expression table}
\usage{
build_expression_table(summary_data, controls, base_group)
}
\description{
Build expression table
}
\keyword{internal}
