% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{collapse_technical_replicates}
\alias{collapse_technical_replicates}
\title{Collapse technical replicates}
\usage{
collapse_technical_replicates(
  raw_data,
  technical_replicates = c("average", "error"),
  technical_sd_threshold = 0.2
)
}
\description{
Collapse technical replicates
}
\keyword{internal}
