% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{compute_y_positions}
\alias{compute_y_positions}
\title{Compute y positions for stat annotations}
\usage{
compute_y_positions(df)
}
\description{
Compute y positions for stat annotations
}
\keyword{internal}
