% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecularBio.R
\docType{package}
\name{moleculaBio}
\alias{molecularBio}
\alias{molecularBio-package}
\alias{moleculaBio}
\title{molecularBio}
\description{
molecularBio is a package of functions useful for the analysis of routine data generated by molecular biology experiments.
}
