% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot.qpcr_analysis}
\alias{plot.qpcr_analysis}
\title{Plot method}
\usage{
\method{plot}{qpcr_analysis}(x, ...)
}
\arguments{
\item{x}{A \code{qpcr_analysis} object.}

\item{...}{Arguments passed to \code{\link[=plot_qPCR]{plot_qPCR()}}.}
}
\value{
Invisibly, the plots returned by \code{\link[=plot_qPCR]{plot_qPCR()}}.
}
\description{
Plot method
}
