% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.qpcr_analysis}
\alias{print.qpcr_analysis}
\title{Print qpcr analysis}
\usage{
\method{print}{qpcr_analysis}(x, ...)
}
\arguments{
\item{x}{A \code{qpcr_analysis} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
\code{x}, invisibly.
}
\description{
Print qpcr analysis
}
