% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.qpcr_stats}
\alias{print.qpcr_stats}
\title{Print statistical results}
\usage{
\method{print}{qpcr_stats}(x, ...)
}
\arguments{
\item{x}{A \code{qpcr_stats} object.}

\item{...}{Additional arguments passed to the next print method.}
}
\value{
\code{x}, invisibly.
}
\description{
Print statistical results
}
