% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{run_gene_statistics}
\alias{run_gene_statistics}
\title{Run statistics for one gene}
\usage{
run_gene_statistics(
  df,
  formula,
  stats_type = c("ttest", "anova"),
  paired = FALSE,
  var.equal = TRUE,
  alternative = "two.sided"
)
}
\description{
Run statistics for one gene
}
\keyword{internal}
