% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{save_plot}
\alias{save_plot}
\title{Save plot helper}
\usage{
save_plot(
  plot,
  filename_base,
  width,
  height,
  dpi,
  save_tiff = TRUE,
  save_svg = TRUE
)
}
\description{
Save plot helper
}
\keyword{internal}
