% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{summary.qpcr_analysis}
\alias{summary.qpcr_analysis}
\title{Summary qpcr analysis}
\usage{
\method{summary}{qpcr_analysis}(object, ...)
}
\arguments{
\item{object}{A \code{qpcr_analysis} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A one-row tibble summarizing the analysis.
}
\description{
Summary qpcr analysis
}
