% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_qPCR.R
\name{validate_qpcr_inputs}
\alias{validate_qpcr_inputs}
\title{Validate qPCR inputs}
\usage{
validate_qpcr_inputs(raw_data, group_label, controls)
}
\description{
Validate qPCR inputs
}
\keyword{internal}
